They move along now
the phrases are moving along. they/re in an array but I need a better way to do this. as they move to their different 'positions' on the belt they need to move to that respective position on the array. Maybe I should do this with literal gameobject positions in the world or something. 1:30am so I'm going to call it a night