partially udating the conveyor
making a better system where I will have an array of phrases and an array of positions that the phrases will be sent through. as each phrase moves to a different spot in the array it will move to the respective position. this work isn't done I really just have the arrays set up. I think that next I need to have them move to their respective position based upon which index of the array they are at.