finished wavecreator
now, based upon my phrasedata class, we generate a list of wave objects which contain quote objects. these quotes contain text as well as an array which tells whether it is good and bad on any given wave. at runtime, we generate the full list of waves with randomly chosen phrases appropriately being good or bad. as each wave is created, it is merely pulled over from the pre-generated wave.